Feat: Seperate useInterval
